Add files via upload
This is the map that I plan to use for a game based on destructible terrain using quad trees. This is a bare bones, early stage in the development of the game. 

Left-click to spawn a set blast on the mouse pointer that will destroy nearby terrain, after a small delay